+++ api/src/application/mod.rs
@@ -1,3 +1,2 @@
 pub mod payment;
-pub mod payment_request;
 pub mod project;
