--- api/src/application/payment_request/mod.rs
@@ -1,1 +1,0 @@
-pub mod create;
