+++ api/src/graphql/context.rs
@@ -6,7 +6,7 @@ use domain::{
 use std::sync::Arc;
 
 pub struct Context {
-	pub create_payment_request_usecase: application::payment_request::create::Usecase,
+	pub create_payment_request_usecase: application::payment::request::Usecase,
 	pub create_payment_usecase: application::payment::create::Usecase,
 	pub create_project_usecase: application::project::create::Usecase,
 	pub assign_project_lead_usecase: application::project::assign_leader::Usecase,
@@ -20,7 +20,7 @@ impl Context {
 		user_repository: Arc<dyn UserRepository>,
 	) -> Self {
 		Self {
-			create_payment_request_usecase: application::payment_request::create::Usecase::new(
+			create_payment_request_usecase: application::payment::request::Usecase::new(
 				uuid_generator.to_owned(),
 				event_publisher.to_owned(),
 				project_repository.to_owned(),
