+++ domain/src/payment/aggregate.rs
@@ -1,6 +1,12 @@
 use super::amount::Amount;
-use crate::{specifications, Aggregate, PaymentEvent, PaymentId, PaymentReceipt, PaymentRequestId};
+#[cfg_attr(test, mockall_double::double)]
+use crate::specifications::{ProjectExists, UserExists};
+use crate::{
+	specifications, Aggregate, AggregateRoot, EventSourcable, PaymentEvent, PaymentId,
+	PaymentReceipt, PaymentRequestId, ProjectId, UserId,
+};
 use serde::{Deserialize, Serialize};
+use serde_json::Value;
 use thiserror::Error;
 
 #[derive(Debug, Default, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -13,13 +19,74 @@ impl Aggregate for Payment {
 	type Id = PaymentId;
 }
 
+impl EventSourcable for Payment {
+	fn apply_event(self, event: &Self::Event) -> Self {
+		match event {
+			PaymentEvent::Requested { id, .. } => Self { id: *id },
+			PaymentEvent::Created { id, .. } => Self { id: *id },
+		}
+	}
+}
+
+impl AggregateRoot for Payment {}
+
 #[derive(Debug, Error)]
 pub enum Error {
+	#[error("Project not found")]
+	ProjectNotFound,
+	#[error("Requestor not found")]
+	RequestorNotFound,
+	#[error("Recipient not found")]
+	RecipientNotFound,
 	#[error(transparent)]
 	Specification(specifications::Error),
 }
 
 impl Payment {
+	#[cfg_attr(feature = "cargo-clippy", allow(clippy::too_many_arguments))]
+	pub async fn request(
+		project_exists_specification: &ProjectExists,
+		user_exists_specification: &UserExists,
+		id: PaymentId,
+		project_id: ProjectId,
+		requestor_id: UserId,
+		recipient_id: UserId,
+		amount_in_usd: u32,
+		reason: Value,
+	) -> Result<Vec<<Self as Aggregate>::Event>, Error> {
+		if !project_exists_specification
+			.is_satisfied_by(&project_id)
+			.map_err(Error::Specification)?
+		{
+			return Err(Error::ProjectNotFound);
+		}
+
+		if !user_exists_specification
+			.is_satisfied_by(&requestor_id)
+			.await
+			.map_err(Error::Specification)?
+		{
+			return Err(Error::RequestorNotFound);
+		}
+
+		if !user_exists_specification
+			.is_satisfied_by(&recipient_id)
+			.await
+			.map_err(Error::Specification)?
+		{
+			return Err(Error::RecipientNotFound);
+		}
+
+		Ok(vec![PaymentEvent::Requested {
+			id,
+			project_id,
+			requestor_id,
+			recipient_id,
+			amount_in_usd,
+			reason,
+		}])
+	}
+
 	pub fn create(
 		id: PaymentId,
 		request_id: PaymentRequestId,
@@ -40,22 +107,65 @@ impl Payment {
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use crate::{BlockchainNetwork, Currency};
-	use rstest::*;
+	#[mockall_double::double]
+	use crate::specifications::ProjectExists;
+	use crate::{BlockchainNetwork, Currency, PaymentRequestId, ProjectId, UserId};
+	use assert_matches::assert_matches;
+	use mockall::predicate::*;
+	use rstest::{fixture, rstest};
 	use rust_decimal_macros::dec;
+	use std::str::FromStr;
 	use testing::fixtures::payment::*;
 	use uuid::Uuid;
 
 	#[fixture]
-	#[once]
 	fn payment_request_id() -> PaymentRequestId {
-		Uuid::new_v4().into()
+		Uuid::from_str("00000000-aaaa-495e-9f4c-038ec0ebecb1").unwrap().into()
 	}
 
 	#[fixture]
-	#[once]
+	fn project_id() -> ProjectId {
+		Uuid::from_str("11111111-aaaa-495e-9f4c-038ec0ebecb1").unwrap().into()
+	}
+
+	#[fixture]
+	fn wrong_project_id() -> ProjectId {
+		Uuid::from_str("11111111-bbbb-495e-9f4c-038ec0ebecb1").unwrap().into()
+	}
+
+	#[fixture]
+	fn requestor_id() -> UserId {
+		Uuid::from_str("22222222-aaaa-495e-9f4c-038ec0ebecb1").unwrap().into()
+	}
+
+	#[fixture]
+	fn wrong_requestor_id() -> UserId {
+		Uuid::from_str("22222222-bbbb-495e-9f4c-038ec0ebecb1").unwrap().into()
+	}
+
+	#[fixture]
+	fn recipient_id() -> UserId {
+		Uuid::from_str("33333333-aaaa-495e-9f4c-038ec0ebecb1").unwrap().into()
+	}
+
+	#[fixture]
+	fn wrong_recipient_id() -> UserId {
+		Uuid::from_str("33333333-bbbb-495e-9f4c-038ec0ebecb1").unwrap().into()
+	}
+
+	#[fixture]
+	fn amount_in_usd() -> u32 {
+		420
+	}
+
+	#[fixture]
+	fn reason() -> Value {
+		serde_json::to_value("{}").unwrap()
+	}
+
+	#[fixture]
 	fn payment_id() -> PaymentId {
-		Uuid::new_v4().into()
+		Uuid::from_str("44444444-aaaa-495e-9f4c-038ec0ebecb1").unwrap().into()
 	}
 
 	#[fixture]
@@ -74,14 +184,14 @@ mod tests {
 
 	#[fixture]
 	fn payment_created_event(
-		payment_request_id: &PaymentRequestId,
-		payment_id: &PaymentId,
+		payment_request_id: PaymentRequestId,
+		payment_id: PaymentId,
 		amount: Amount,
 		receipt: PaymentReceipt,
 	) -> PaymentEvent {
 		PaymentEvent::Created {
-			id: *payment_id,
-			request_id: *payment_request_id,
+			id: payment_id,
+			request_id: payment_request_id,
 			amount,
 			receipt,
 		}
@@ -89,15 +199,202 @@ mod tests {
 
 	#[rstest]
 	fn test_create(
-		payment_request_id: &PaymentRequestId,
-		payment_id: &PaymentId,
+		payment_request_id: PaymentRequestId,
+		payment_id: PaymentId,
 		payment_created_event: PaymentEvent,
 		amount: Amount,
 		receipt: PaymentReceipt,
 	) {
-		let events = Payment::create(*payment_id, *payment_request_id, amount, receipt)
+		let events = Payment::create(payment_id, payment_request_id, amount, receipt)
 			.expect("Problem when creating payment");
 
 		assert_eq!(events, vec![payment_created_event]);
+	}
+
+	#[rstest]
+	async fn test_request(
+		payment_id: PaymentId,
+		project_id: ProjectId,
+		requestor_id: UserId,
+		recipient_id: UserId,
+		amount_in_usd: u32,
+		reason: Value,
+	) {
+		let mut project_exists_specification = ProjectExists::default();
+		project_exists_specification
+			.expect_is_satisfied_by()
+			.with(eq(project_id))
+			.once()
+			.returning(|_| Ok(true));
+
+		let mut user_exists_specification = UserExists::default();
+		user_exists_specification
+			.expect_is_satisfied_by()
+			.with(eq(requestor_id))
+			.once()
+			.returning(|_| Ok(true));
+		user_exists_specification
+			.expect_is_satisfied_by()
+			.with(eq(recipient_id))
+			.once()
+			.returning(|_| Ok(true));
+
+		let events = Payment::request(
+			&project_exists_specification,
+			&user_exists_specification,
+			payment_id,
+			project_id,
+			requestor_id,
+			recipient_id,
+			amount_in_usd,
+			reason.clone(),
+		)
+		.await
+		.unwrap();
+
+		assert_eq!(events.len(), 1);
+		assert_eq!(
+			events[0],
+			PaymentEvent::Requested {
+				id: payment_id,
+				project_id,
+				requestor_id,
+				recipient_id,
+				amount_in_usd,
+				reason,
+			}
+		);
+	}
+
+	#[rstest]
+	async fn test_request_with_wrong_project_id(
+		payment_id: PaymentId,
+		wrong_project_id: ProjectId,
+		requestor_id: UserId,
+		recipient_id: UserId,
+		amount_in_usd: u32,
+		reason: Value,
+	) {
+		let mut project_exists_specification = ProjectExists::default();
+		project_exists_specification
+			.expect_is_satisfied_by()
+			.with(eq(wrong_project_id))
+			.once()
+			.returning(|_| Ok(false));
+
+		let user_exists_specification = UserExists::default();
+
+		let result = Payment::request(
+			&project_exists_specification,
+			&user_exists_specification,
+			payment_id,
+			wrong_project_id,
+			requestor_id,
+			recipient_id,
+			amount_in_usd,
+			reason,
+		)
+		.await;
+
+		assert!(result.is_err());
+		assert_matches!(result, Err(Error::ProjectNotFound));
+	}
+
+	#[rstest]
+	async fn test_request_with_wrong_requestor_id(
+		payment_id: PaymentId,
+		project_id: ProjectId,
+		wrong_requestor_id: UserId,
+		recipient_id: UserId,
+		amount_in_usd: u32,
+		reason: Value,
+	) {
+		let mut project_exists_specification = ProjectExists::default();
+		project_exists_specification
+			.expect_is_satisfied_by()
+			.with(eq(project_id))
+			.once()
+			.returning(|_| Ok(true));
+
+		let mut user_exists_specification = UserExists::default();
+		user_exists_specification
+			.expect_is_satisfied_by()
+			.with(eq(wrong_requestor_id))
+			.once()
+			.returning(|_| Ok(false));
+
+		let result = Payment::request(
+			&project_exists_specification,
+			&user_exists_specification,
+			payment_id,
+			project_id,
+			wrong_requestor_id,
+			recipient_id,
+			amount_in_usd,
+			reason.clone(),
+		)
+		.await;
+
+		assert!(result.is_err());
+		assert_matches!(result, Err(Error::RequestorNotFound));
+	}
+
+	#[rstest]
+	async fn test_request_with_wrong_recipient_id(
+		payment_id: PaymentId,
+		project_id: ProjectId,
+		requestor_id: UserId,
+		wrong_recipient_id: UserId,
+		amount_in_usd: u32,
+		reason: Value,
+	) {
+		let mut project_exists_specification = ProjectExists::default();
+		project_exists_specification
+			.expect_is_satisfied_by()
+			.with(eq(project_id))
+			.once()
+			.returning(|_| Ok(true));
+
+		let mut user_exists_specification = UserExists::default();
+		user_exists_specification
+			.expect_is_satisfied_by()
+			.with(eq(requestor_id))
+			.once()
+			.returning(|_| Ok(true));
+		user_exists_specification
+			.expect_is_satisfied_by()
+			.with(eq(wrong_recipient_id))
+			.once()
+			.returning(|_| Ok(false));
+
+		let result = Payment::request(
+			&project_exists_specification,
+			&user_exists_specification,
+			payment_id,
+			project_id,
+			requestor_id,
+			wrong_recipient_id,
+			amount_in_usd,
+			reason.clone(),
+		)
+		.await;
+
+		assert!(result.is_err());
+		assert_matches!(result, Err(Error::RecipientNotFound));
+	}
+
+	#[rstest]
+	fn test_event_sourced(payment_id: PaymentId) {
+		let event = PaymentEvent::Requested {
+			id: payment_id,
+			project_id: Default::default(),
+			requestor_id: Default::default(),
+			recipient_id: Default::default(),
+			amount_in_usd: Default::default(),
+			reason: Default::default(),
+		};
+
+		let payment = Payment::from_events(&[event]);
+		assert_eq!(payment.id, payment_id);
 	}
 }
