+++ event-listeners/src/domain/projectors/payment_request.rs
@@ -1,7 +1,7 @@
 use crate::domain::{EventListener, PaymentRequest, ProjectionRepository};
 use anyhow::Result;
 use async_trait::async_trait;
-use domain::{Event, PaymentRequestEvent};
+use domain::{Event, PaymentEvent};
 use std::sync::Arc;
 
 pub struct Projector {
@@ -17,24 +17,23 @@ impl Projector {
 #[async_trait]
 impl EventListener for Projector {
 	async fn on_event(&self, event: &Event) -> Result<()> {
-		if let Event::PaymentRequest(event) = event {
-			match event {
-				PaymentRequestEvent::Created {
-					id,
-					project_id,
-					requestor_id,
-					recipient_id,
-					amount_in_usd,
-					reason,
-				} => self.repository.insert(&PaymentRequest::new(
-					(*id).into(),
-					(*project_id).into(),
-					(*requestor_id).into(),
-					(*recipient_id).into(),
-					*amount_in_usd as i64,
-					reason.clone(),
-				))?,
-			}
+		if let Event::Payment(PaymentEvent::Requested {
+			id,
+			project_id,
+			requestor_id,
+			recipient_id,
+			amount_in_usd,
+			reason,
+		}) = event
+		{
+			self.repository.insert(&PaymentRequest::new(
+				(*id).into(),
+				(*project_id).into(),
+				(*requestor_id).into(),
+				(*recipient_id).into(),
+				*amount_in_usd as i64,
+				reason.clone(),
+			))?
 		}
 		Ok(())
 	}
