+++ event-store/src/lib.rs
@@ -64,15 +64,14 @@ impl IdentifiableAggregate for Event {
 	fn aggregate_id(&self) -> String {
 		match &self {
 			Event::Project(event) => match event {
-				backend_domain::ProjectEvent::Created { id, .. } => id.to_string(),
-				backend_domain::ProjectEvent::LeaderAssigned { id, .. } => id.to_string(),
+				backend_domain::ProjectEvent::Created { id, .. }
+				| backend_domain::ProjectEvent::LeaderAssigned { id, .. } => id.to_string(),
 			},
 			Event::Payment(event) => match event {
-				backend_domain::PaymentEvent::Created { id, .. } => id.to_string(),
+				backend_domain::PaymentEvent::Requested { id, .. }
+				| backend_domain::PaymentEvent::Created { id, .. } => id.to_string(),
 			},
-			Event::PaymentRequest(event) => match event {
-				backend_domain::PaymentRequestEvent::Created { id, .. } => id.to_string(),
-			},
+			Event::PaymentRequest(_) => unimplemented!(),
 		}
 	}
 }
